Fixed bug involving order of learners.
This commit addresses a bug where the learners would be sorted in a different order than they were read from the file. Ideally, this should not matter, since all learners put in their bid and the bid is used to evaluate. Problems arose, however, when the difference between the winning bid and runner-up(s) was less than the value set for macro BID_EPSILON. I am ashamed of this commit for it is nothing more than a hack.